Replaced optional protocol methods with protocol extensions with empty implementations
Added schemes to build libraries for iOS and tvOS.

Replaced static libraries with frameworks for carthage support

Removed invalid characters from framework names

Removed iOS framework as it isn't building with carthage.  The error message seems to indicate something about not enabling bit code which is required for tvOS but as it is an iOS framework it shouldn't be needed...

Enable bit code???  I switched the base sdk to iOS then enabled bit code.